Init should take a list of peers
Addressed reviews
- Returning errors in SavePeerstore when they happen (I think they are
warnings now because that function was only used during cluster
shutdown which meant it was a best effort thing and in some cases like
tests saving the peerstore was expected to fail).
- Erroring when the addrInfos have nothing but a peer id
(init -peers /ipfs/Qm... works and it should error because there is no
addresses associated to the given peer ID)
- A sharness test checking the peerstore file was written with the
expected contents.