Fix #949: Cluster peers should advertise their external earlier
This lowers the activation threshold for an observed addressed
to 1 and increases the OwnObservedAddressTTL to infinite.

This should let cluster peers (particularly in small clusters)
advertise their known external addresses asap so that other peers
can include them in their peerstores (and save them on shutdown).

By default this only happens when an external address has received 4
connections in the last 40 minutes. In NATed environments this may
result in cluster peers advertising right away their NAT-translated address/ports
which may not be dialable, but in other cases that address seems to be
indeed dialable (docker).