Set the RPC Protocol Version to a fixed value
The RPC endpoints no longer change much, but every minor release enforces that
all peers in a Cluster must run on the same version.

This was meant to allow greater flexibility when upgrading the RPC
endpoints. Now we should not be in a situation that releasing a minor cluster
version causes peers incompatibilities, even though nothing changes on the RPC
side of things.

Therefore we make 0.12 the RPC protocol version going forward and detach it
from the actual cluster version from this point.