pstoremgr: bootstrap: harden reconnection logic
This commit lets peers reconnect to one of the "peer_addresses" every 5
minutes. It is assumed that these are addresses for peers that are likely
honest. In a very large cluster with many untrusted peers we want peers to
check-in with one peer_address every now and then, as it is more likely that
it is correctly relaying pubsub messages etc., which we cannot ensure for
other random peers that may be part of the cluster.

In a sense, peer_addresses becomes more like the notion of bootstrap
nodes. They get Permanent Address TTL in the peerstore and are given higher
priority than peerstore addresses.

We additionally make sure that peers with the same priority are sorted in
random orders so that load is distributed among available peer_addresses.